Fixes url publisher toggle
Currently, the publisher toggle only functions as a switch for the pinned status of a publisher if it is already in the ledger table. This fix adds an app action so that a publisher may be manually included and pinned at any time via the toggle button, in spite of the auto include setting and minimum duration and visit settings.